requirements.txt
ttkbootstrap>=1.10.1
Pillow>=9.0.0
fpdf2>=2.7.8 